Refactor slack and burndown functions into their own files
Fix most clippy errors
