+++ card-counter/lambda/src/burndown_lambda.rs
@@ -1,146 +1,22 @@
-use card_counter::{commands::burndown::{self, BurndownOptions}, database::{Database, DateRange, aws::Aws, config::{Config, trello_auth_from_env}}, errors::*, kanban::{self, Kanban, trello::{TrelloAuth, TrelloClient}}};
+mod burndown_helpers;
+mod slack_helpers;
+use burndown_helpers::*;
+use slack_helpers::*;
 
-use std::{collections::HashMap, str::FromStr, string::ParseError};
+use card_counter::errors::*;
+
+use std::{collections::HashMap, str::FromStr};
 
 use rusoto_core::Region;
 use rusoto_s3::{PutObjectRequest, S3Client, S3};
-
-#[macro_use]
-use serde::{Deserialize, Serialize};
 use aws_lambda_events::encodings::Body;
 use aws_lambda_events::event::apigw::{ApiGatewayProxyRequest, ApiGatewayProxyResponse};
-use http::header::{HeaderMap, CONTENT_TYPE};
 use lambda::{handler_fn, Context};
-use serde_urlencoded;
+use http::header::{HeaderMap, CONTENT_TYPE};
 
-#[macro_use]
+
 use log::{error, info};
-use simple_logger;
 
-#[derive(Debug, Deserialize, Clone)]
-struct SlackCommand {
-  token: Option<String>,
-  team_id: Option<String>,
-  team_domain: Option<String>,
-  channel_id: Option<String>,
-  channel_name: Option<String>,
-  user_id: Option<String>,
-  user_name: Option<String>,
-  command: Option<String>,
-  text: String,
-  api_app_id: Option<String>,
-  is_enterprise_install: bool,
-  response_url: Option<String>,
-  trigger_id: Option<String>,
-}
-
-#[derive(Debug, Serialize)]
-struct SlackBlock {
-  blocks: Vec<SlackMessage>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  response_type: Option<String>
-}
-#[derive(Debug, Serialize, Default)]
-struct SlackMessage {
-  #[serde(rename = "type")]
-  slack_type: String,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  elements: Option<Vec<HashMap<String, String>>>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  text: Option<HashMap<String, String>>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  title: Option<HashMap<String, String>>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  image_url: Option<String>,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  alt_text: Option<String>,
-}
-
-fn context_error(message: String) -> SlackMessage {
-  let mut context: HashMap<String, String> = HashMap::new();
-  context.insert("type".to_string(), "mrkdwn".to_string());
-  context.insert(
-    "text".to_string(),
-    format!(
-      "I'm sorry, I didn't understand your request. Try formatting your request like\n{}",
-      message
-    ),
-  );
-  SlackMessage {
-    slack_type: "context".to_string(),
-    elements: Some(vec![context]),
-    text: None,
-    ..SlackMessage::default()
-  }
-}
-#[derive(Debug, PartialEq)]
-struct BurndownConfig {
-  pub start: Option<String>,
-  pub end: Option<String>,
-  pub board_id: Option<String>,
-}
-impl BurndownConfig {
-  fn helper_string(&self) -> Option<String> {
-    if self.start.is_none() || self.end.is_none() || self.board_id.is_none() {
-      Some(format!(
-        "/card-counter burndown from {} to {} for {}",
-        self.start.as_ref().unwrap_or(&"YYYY-MM-DD".to_string()),
-        self.end.as_ref().unwrap_or(&"YYYY-MM-DD".to_string()),
-        self.board_id.as_ref().unwrap_or(&"<board-id>".to_string())
-      ))
-    } else {
-      None
-    }
-  }
-}
-
-impl Default for BurndownConfig {
-  fn default() -> Self {
-    Self {
-      start: None,
-      end: None,
-      board_id: None,
-    }
-  }
-}
-
-impl FromStr for BurndownConfig {
-  type Err = ParseError;
-  fn from_str(s: &str) -> Result<Self, Self::Err> {
-    let mut config = BurndownConfig::default();
-    let tokens: Vec<&str> = s.trim().split(" ").collect();
-    let mut i = 0;
-
-    while i < tokens.len() {
-      if tokens[i].to_lowercase() == "from" && i + 1 < tokens.len() {
-        config.start = Some(tokens[i + 1].to_string());
-      } else if tokens[i].to_lowercase() == "to" && i + 1 < tokens.len() {
-        config.end = Some(tokens[i + 1].to_string());
-      } else if tokens[i].to_lowercase() == "for" && i + 1 < tokens.len() {
-        config.board_id = Some(tokens[i + 1].to_string());
-      }
-      i = i + 1;
-    }
-    Ok(config)
-  }
-}
-
-// Often times a user will use the boards shortLink, this is an 8
-// character string, but we store the index in dynamodb as the board's
-// full id, a 24 character string. So we need to make sure we have the
-// full id to work.
-async fn get_full_board_id(board_id: String) -> Result<String> {
-  let client = TrelloClient {
-    client: reqwest::Client::new(),
-    auth: trello_auth_from_env().unwrap()
-  };
-
-  if board_id.len() == 24 {
-    Ok(board_id)
-  } else {
-    Ok(client.get_board(&board_id).await?.id)
-  }
-}
 
 type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
 
@@ -155,9 +31,10 @@ async fn main() -> Result<(), Error> {
   lambda::run(func).await?;
   Ok(())
 }
+
 async fn lambda_apigw_wrapper(
   api_event: ApiGatewayProxyRequest,
-  context: Context,
+  _context: Context,
 ) -> Result<ApiGatewayProxyResponse> {
   info!("{:?}", api_event);
   let event: SlackCommand = serde_urlencoded::from_str(&api_event.body.unwrap())?;
@@ -168,6 +45,20 @@ async fn lambda_apigw_wrapper(
   info!("{:?}", apigw_response);
   Ok(apigw_response)
 }
+
+fn default_gateway_response(body: SlackBlock) -> ApiGatewayProxyResponse {
+  let mut headers = HeaderMap::new();
+  headers.insert(CONTENT_TYPE, "application/json".parse().unwrap());
+
+  ApiGatewayProxyResponse {
+    status_code: 200,
+    multi_value_headers: HeaderMap::new(),
+    headers,
+    body: Some(Body::Text(serde_json::json!(&body).to_string())),
+    is_base64_encoded: Some(false),
+  }
+}
+
 /// you can invoke the lambda with a JSON payload, which is parsed using the CustomEvent struct.
 async fn my_handler(event: SlackCommand) -> Result<SlackBlock> {
   let config = match BurndownConfig::from_str(&event.text) {
@@ -224,23 +115,6 @@ async fn my_handler(event: SlackCommand) -> Result<SlackBlock> {
   };
   Ok(block)
 }
-
-async fn generate_burndown_chart(start: &str, end: &str, board_id: &str) -> eyre::Result<String> {
-  let client: Box<dyn Database> = Box::new(Aws::init(&Config::default()).await?);
-
-  let range = DateRange::from_strs(start, end);
-  let options = BurndownOptions {
-    board_id: board_id.to_string(),
-    range,
-    client,
-    filter: Some("NoBurn".into()),
-  };
-  info!("{:?}", options.board_id);
-  info!("{:?}", options.range);
-  let burndown = options.into_burndown().await?;
-  burndown.as_svg()
-}
-
 async fn upload_chart_to_s3(chart: &str, bucket: &str, date_range: &str) -> Result<()> {
   let client = S3Client::new(Region::default());
 
@@ -257,45 +131,4 @@ async fn upload_chart_to_s3(chart: &str, bucket: &str, date_range: &str) -> Resu
   info!("{:?}", result);
 
   Ok(())
-}
-
-fn validate_env_vars() -> Result<()> {
-  if std::env::var("BUCKET_NAME").is_err() {
-    panic!("Unable to find env variable BUCKET_NAME");
-  }
-  Ok(())
-}
-
-fn default_gateway_response(body: SlackBlock) -> ApiGatewayProxyResponse {
-  let mut headers = HeaderMap::new();
-  headers.insert(CONTENT_TYPE, "application/json".parse().unwrap());
-
-  ApiGatewayProxyResponse {
-    status_code: 200,
-    multi_value_headers: HeaderMap::new(),
-    headers,
-    body: Some(Body::Text(serde_json::json!(&body).to_string())),
-    is_base64_encoded: Some(false),
-  }
-}
-
-#[cfg(test)]
-mod test {
-  use std::str::FromStr;
-
-  use crate::BurndownConfig;
-
-  #[test]
-  fn it_makes_a_burndown_cfg() {
-    let result =
-      BurndownConfig::from_str("burndown from 2020-01-01 to 2020-10-01 for 3em95wSl").unwrap();
-    assert_eq!(
-      result,
-      BurndownConfig {
-        start: Some("2020-01-01".to_string()),
-        end: Some("2020-10-01".to_string()),
-        board_id: Some("3em95wSl".to_string())
-      }
-    );
-  }
 }
